requirements.txt
Langchain
ipykernel
python-dotenv
Langchain_community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
lxml
langchain-openai
chromadb
sentence-transformers
langchain_huggingface
faiss-cpu
langchain_chroma
streamlit
langchain_groq
fastapi
uvicorn
langserve
sse_starlette
duckdb
duckduckgo-search